refactor: extract UpdateCoordinator from App.xaml.cs
Moves update check, download, install, and dialog logic from App.xaml.cs
into a dedicated UpdateCoordinator service. App retains a single
_updateCoordinator field and delegates all four call sites (startup,
CLI arg, IAppCommands, deep link).

XamlRoot is resolved via Func<XamlRoot?> (hub-first, keepalive fallback)
so the coordinator has no direct window reference. Fixes a latent bug
where UpdateCommandCenterInfo.Detail was mutated in-place on the existing
object instead of being replaced — AppState.UpdateInfo only fires
PropertyChanged on assignment, so the in-place mutation was silently
dropped by the UI.

Addresses part of #554.
